Move dialogs to `OfficeApi.Extensions` project
