+++ Source/NetOffice.sln
@@ -45,6 +45,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MSDATASRCApi", "MSDATASRC\M
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NetOffice.Tests", "NetOffice.Tests\NetOffice.Tests.csproj", "{67372CD1-E220-44DA-AB36-F489A593923C}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "OfficeApi.Extensions", "Office.Extensions\OfficeApi.Extensions.csproj", "{A928079E-A92E-4964-A586-EFED61ED665A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -112,6 +114,10 @@ Global
 		{67372CD1-E220-44DA-AB36-F489A593923C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{67372CD1-E220-44DA-AB36-F489A593923C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{67372CD1-E220-44DA-AB36-F489A593923C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A928079E-A92E-4964-A586-EFED61ED665A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A928079E-A92E-4964-A586-EFED61ED665A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A928079E-A92E-4964-A586-EFED61ED665A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A928079E-A92E-4964-A586-EFED61ED665A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
