+++ Source/Office.Extensions/Tools/Dialogs/AboutDialog.1033.xml
