+++ Source/Office.Extensions/Tools/Dialogs/AboutDialog.resx
