+++ Source/Office.Extensions/Tools/Dialogs/DiagnosticsDialog.1033.xml
