+++ Source/Office.Extensions/Tools/Dialogs/DiagnosticsDialog.resx
