+++ Source/Office.Extensions/Tools/Dialogs/DialogLocalizationSettings.cs
