+++ Source/Office.Extensions/Tools/Dialogs/ErrorDialog.cs
@@ -5,6 +5,7 @@ using System.Data;
 using System.Drawing;
 using System.Text;
 using System.Windows.Forms;
+using NetOffice.OfficeApi.Tools.Contribution;
 
 namespace NetOffice.OfficeApi.Tools.Dialogs
 {
