+++ Source/Office.Extensions/Tools/Dialogs/ErrorDialog.resx
