+++ Source/Office.Extensions/Tools/Dialogs/PropertyGridDialog.cs
