+++ Source/Office.Extensions/Tools/Dialogs/PropertyGridDialog.resx
