+++ Source/Office.Extensions/Tools/Dialogs/RichTextDialog.1033.xml
