+++ Source/Office.Extensions/Tools/Dialogs/RichTextDialog.cs
