+++ Source/Office.Extensions/Tools/Dialogs/RichTextDialog.resx
