+++ Source/Office.Extensions/Tools/Dialogs/ToolsDialog.resx
